PDFBOX-2590: Simplify marked content API
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1649990 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/cos/COSDocument.java
@@ -261,14 +261,14 @@ public class COSDocument extends COSBase implements Closeable
      *
      * @param versionValue The version of the PDF document.
      */
-    public void setVersion( float versionValue )
+    public void setVersion( double versionValue )
     {
         // update header string
         if (versionValue != version) 
         {
             headerString = headerString.replaceFirst(String.valueOf(version), String.valueOf(versionValue));
         }
-        version = versionValue;
+        version = (float)versionValue;
     }
 
     /**
